[Feat] #168 ReviewListView UI 작업
